backend/requirements.txt
# FastAPI Framework
fastapi>=0.115.0

# ASGI Server
uvicorn[standard]>=0.32.0

# Database & ORM
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9

# Data Validation (compatible with Python 3.13)
pydantic>=2.10.0
pydantic[email]>=2.10.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.0,<5.0.0
python-multipart>=0.0.6

# Environment Variables (optional but recommended)
python-dotenv>=1.0.0

